Extract distinct code into separate modules; reorganize code to support isolated testing
