+++ ops/terraform/services/eft/lambda_src/sftp_outbound_transfer/.gitignore
@@ -1,0 +1,1 @@
+__pycache__/*
