+++ ops/terraform/services/eft/lambda_src/sftp_outbound_transfer/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+pythonpath = .
+testpaths = tests
