+++ ops/terraform/services/eft/lambda_src/sftp_outbound_transfer/tests/.gitignore
@@ -1,0 +1,1 @@
+__pycache__/*
